+++ alicloud/import_alicloud_pvtz_zone_test.go
@@ -1,0 +1,28 @@
+package alicloud
+
+import (
+	"testing"
+
+	"github.com/hashicorp/terraform/helper/resource"
+)
+
+func TestAccAlicloudPvtzZone_importBasic(t *testing.T) {
+	resourceName := "alicloud_pvtz_zone.foo"
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccAlicloudPvtzZoneDestroy,
+		Steps: []resource.TestStep{
+			resource.TestStep{
+				Config: testAccPvtzZoneConfig,
+			},
+
+			resource.TestStep{
+				ResourceName:      resourceName,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
+		},
+	})
+}